Support sending alerts to multiple SMS numbers
This is done by creating a new struct based on sendOptions but with identical
data. This is required to avoid the same data structure in memory getting
modified in the for loop before SMS messages are sent - end result of which
would be sending all the SMS messages to the last receiver on the list.